chore(core): rename IDataLoader to IDataStorageProvider
