+++ packages/core/src/databus/databus.ts
@@ -16,10 +16,9 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { IDataLoader } from './data.loader.interface';
-import { IDataSaver } from './data.saver.interface';
-import { Database } from './database';
-import { IStoreProvider } from './store.provider.interface';
+import { Database } from './logic';
+import { IDataStorageProvider } from './providers';
+import { IStoreProvider } from './providers';
 
 /**
  * The entry point of the DataBus mechanism.
@@ -35,12 +34,11 @@ export class DataBus {
    * Create a DataBus instance.
    */
   public static create(options: IDataBusInitOptions): DataBus {
-    const { dataLoader, dataSaver, storeProvider } = options;
+    const { dataStorageProvider, storeProvider } = options;
     const databus = new DataBus();
 
     const db = databus.getDatabase();
-    db.setDataLoader(dataLoader);
-    db.setDataSaver(dataSaver);
+    db.setDataStorageProvider(dataStorageProvider);
     db.setStoreProvider(storeProvider);
 
     return databus;
@@ -48,10 +46,10 @@ export class DataBus {
 
   /**
    * Get the database of a DataBus instance.
-   * 
+   *
    * Currently, only one database exists for a DataBus instance. In the future, there may be multiple databases
    * for a DataBus instance. And the use case is:
-   * 
+   *
    * - For frontend, there is only one user, so one database suffice.
    * - For backend, the server creates many databases, one database corresponds to one user respectively.
    */
@@ -64,7 +62,6 @@ export class DataBus {
  * The options to intiailize a DataBus instance.
  */
 export interface IDataBusInitOptions {
-  dataLoader: IDataLoader;
-  dataSaver: IDataSaver;
+  dataStorageProvider: IDataStorageProvider;
   storeProvider: IStoreProvider;
 }
