+++ packages/core/src/databus/providers/data.storage.provider.interface.ts
@@ -1,3 +1,4 @@
+
 /**
  * APITable <https://github.com/apitable/apitable>
  * Copyright (C) 2022 APITable Ltd. <https://apitable.com>
@@ -16,15 +17,21 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { IBaseDatasheetPack } from 'exports/store';
+import { Datasheet } from 'databus/logic';
+import { IBaseDatasheetPack, IReduxState } from 'exports/store';
+import { Store } from 'redux';
+import { IDataSaver } from './data.saver.interface';
 
 /**
- * A data loader is responsible for loading internal datasheet packs for `Datasheet`s from different data sources.
+ * A data storage provider is responsible for loading internal datasheet packs for `Datasheet`s from different data sources, as well as 
+ * saving the results of executing commands into various data storage system.
  *
+ * For example, in the front end, a data storage provider implementation fetches datasheet packs from the server, and sends command results
+ * to the server; while in the back end, another data storage provider implementation loads datasheet packs from the 
  * For example, in the front end, a data loader implementation fetches datasheet packs from the server; while in the back end,
- * another data loader implementation loads datasheet packs from the database.
+ * another data loader implementation loads datasheet packs from the database, and saves the command execution result into the database.
  */
-export interface IDataLoader {
+export interface IDataStorageProvider extends IDataSaver {
   /**
    * Loads a datasheet pack for a datasheet from the data source. Returns null if the datasheet is not found.
    *
@@ -38,3 +45,11 @@ export interface IDataLoader {
  */
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface ILoadDatasheetPackOptions {}
+
+/**
+ * The options of saving command execution results. Implementors of `IDataSaver` can derive this interface, adding necessary fields.
+ */
+export interface ISaveOpsOptions {
+  datasheet: Datasheet;
+  store: Store<IReduxState>;
+}
